requirements.txt
Flask>=2.0
Flask-Session
google-auth
google-auth-oauthlib
google-api-python-client
scikit-learn
pandas
nltk
joblib
flask-cors
Flask-Limiter
